move optimizations back to typesystem file
